% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative.R
\name{AIC}
\alias{AIC}
\title{Akaike Information Criterion for PPCA.}
\usage{
AIC(lambda, M)
}
\arguments{
\item{lambda}{a numeric vector of positive sample eigenvalues that sums to N}

\item{M}{a positive integer for the number of observations or features``}
}
\value{
an integer K between 1 and N that minimizes the AIC.
}
\description{
The function returns the dimension that minimized the AIC based on the
   profile log-likelihood while considering all possible dimensions.
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale((X))))$mat))$val
AIC(lambda = eigen_values, M = 1000)
AIC(lambda = eigen_values, M = 5000)
}

}
\references{
Akaike, H. (1974), A new look at the statistical model identification,
   \emph{IEEE Transactions on Automatic Control}, \strong{19} (6): 716–723, MR 0423716,
   doi:10.1109/TAC.1974.1100705.
}
\keyword{criterion,}
\keyword{information}
\keyword{log-likelihood,}
\keyword{model}
\keyword{profile}
\keyword{selection}
