% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bai2002.R
\name{Bai2002}
\alias{Bai2002}
\title{A bias corrected criterion for selecting number of principal components}
\usage{
Bai2002(x = NULL, lambda = NULL, M = NULL, constant = 0.05,
  choice = 1, unbiased = T)
}
\arguments{
\item{x}{a data matrix with the number of rows to be reduced; only complete columns are used.}

\item{lambda}{a numeric vector of sample eigenvalues of the covariance matrix of t(\code{x})}

\item{M}{if \code{x} were not supplied, \code{M} should be given as the number of columns of \code{x}.}

\item{constant}{a small prefixed constant and set to the recommended value of 0.05. See Bai and Ng (2002s) for details.}

\item{choice}{an integer specifying the choice of the preferred asymptotics. See Bai and Ng (2002s) for details.}

\item{unbiased}{a logical specifying whether the residual variance estimate used in the criterion should be biased or unbiased.
See Bai and Ng (2002) for details.}
}
\value{
an integer K
}
\description{
The function returns the choice for PCA as a by-product of the bias-corrected residual variance estimate.
}
\examples{
\dontrun{
X <- MASS::mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(X)))$mat))$val
Bai2002(lambda = eigen_values, M = 100)
Bai2002(lambda = eigen_values, M = 5000)
}

}
\references{
Bai, Jushan, and Serena Ng (2002). Determining the number of factors in approximate factor models.
   \emph{Econometrica} \strong{70.1}: 191-221.
}
\keyword{Bayesian}
\keyword{Laplace's}
\keyword{criterion,}
\keyword{evidence}
\keyword{information}
\keyword{log-likelihood,}
\keyword{method,}
\keyword{model}
\keyword{profile}
\keyword{selection,}
