% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bai2002.R
\name{BaiNg}
\alias{BaiNg}
\title{Bai and Ng's Criteria for Selecting Number of Probabilistic Factors}
\usage{
BaiNg(lambda, M, bias = TRUE, option = 1, tau = 0.001, verbose = FALSE)
}
\arguments{
\item{lambda}{a numeric vector of sample eigenvalues of length $n$.}

\item{M}{the number of observations.}

\item{bias}{a logical specifying whether the residual variance estimate used in the criterion should be biased or unbiased. See Bai and Ng (2002) for details.}

\item{option}{an integer specifying the choice of the preferred asymptotics regime. See Bai and Ng (2002) for details.}

\item{tau}{a tolerance threshold for the smallest eigenvalue, the default value is 0.001.}

\item{verbose}{a logical specifying whether the posterior evidence or
the integer that minimized the evidence should be returned}
}
\value{
an integer K between 1 and n
}
\description{
The function returns the choice for the number of factors in the context of probabilistic factor analysis.
}
\examples{
\dontrun{
X <- MASS::mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(X)))$mat))$val
BaiNg(lambda = eigen_values, M=100)
BaiNg(lambda = eigen_values, M=5000)
}

}
\references{
Bai, J., & Ng, S. (2002). Determining the number of factors in approximate factor models. **Econometrica**, *70*(1), 191-221. <doi:10.1111/1468-0262.00273>
}
\keyword{Bayesian}
\keyword{Laplace's}
\keyword{criterion,}
\keyword{evidence}
\keyword{information}
\keyword{log-likelihood,}
\keyword{method,}
\keyword{model}
\keyword{profile}
\keyword{selection,}
