% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative.R
\name{elbowEigen}
\alias{elbowEigen}
\title{Alternative methods based on the detection of elbow in sample eigenvalues.}
\usage{
elbowEigen(lambda, methods = c("adjD", "cumD", "varD", "cumlog", "logsigma2"))
}
\arguments{
\item{lambda}{a numeric vector of positive sample eigenvalues that sums to $N$.}

\item{methods}{a character vector of possible methods to detect an elbow in sample eigenvalues.}
}
\value{
an integer $K$ between 1 and $N$.
}
\description{
The function returns the dimension that is selected based on various methods that
   attempt to detect an ``elbow'' in sample eigenvalues.
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(X)))$mat))$val
elbowEigen(lambda = eigen_values)
elbowEigen(lambda = eigen_values)
}

}
\author{
Wei Q. Deng, \email{deng@utstat.toronto.edu}
}
\keyword{approach}
\keyword{eigenvalues,}
\keyword{elbow}
\keyword{sample}
