% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_singular.R
\name{get_data_singular}
\alias{get_data_singular}
\title{Simulate Data from Eigenvalue Structure}
\usage{
get_data_singular(N, K, M, sq_singular = NULL, sigma2 = NULL, last = NULL,
  trend = NULL, rho = NULL, df = NULL, dist = "norm", datamat = TRUE)
}
\arguments{
\item{N}{the full dimension of the data}

\item{K}{the true dimension of the}

\item{M}{the number of features of observations}

\item{sq_singular}{a vector of numeric values for the squared singular values.
The other parameters can be skipped if this is supplied along with \eqn{N, K, M}.}

\item{sigma2}{a positive numeric between 0 and 1 for the error variance.}

\item{last}{a positive numeric within reasonable range for the difference
between the Kth eigenvalue and the (\eqn{K}+1)th, a very large difference
might not be possible if \eqn{K} is large.}

\item{trend}{a character, one of \code{exponential}, \code{linear} and \code{quadratic}
for the type of trend in squared singular values}

\item{rho}{a numeric value between 0 and 1 for the amount of auto-correlation, i.e. correlation
between sequential observations or features.}

\item{df}{an integer for the degrees of freedom if \code{dist} == \code{t}}

\item{dist}{a character specifying the error distribution to be one of \code{norm}, \code{t},
for normal and student's t-distribution, respectively.}

\item{datamat}{a logical to indicate whether both the data matrix and
sample eigenvalues or only the sample eigenvalues should be returned}
}
\value{
a list containing the simulated data matrix and
   sample eigenvalues or a numerical vector of sample eigenvalues.
}
\description{
The function returns either the results of penalized profile
   log-likelihood given a matrix of data or a vector of sample
   eigenvalues. The data matrix has the following decomposition
   \eqn{ X = WL + error}, where the rows of \eqn{X} are linear
   projections onto the subspace \eqn{ W } by some arbitrary latent
   vector plus error. The solution finds the rank of \eqn{W}, which
   represents the hidden structure in the data, such that \eqn{ X-WL }
   have independent and identically distributed components.
}
\examples{
\dontrun{
get_data_singular(N = 200, K = 5, M = 1000, sq_singular = c(5,4,2,1,1))
get_data_singular(N = 200, K = 5, M = 1000, sigma2 = 0.2, last= 0.1, trend = "exponential")
get_data_singular(N = 200, K = 5, M = 1000, sigma2 = 0.8, last= 0.1, trend = "exponential",
   rho = 0.2, df = 5, dist = "t")
}
}
\author{
Wei Q. Deng, \email{deng@utstat.toronto.edu}
}
