% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative.R
\name{INFO_ppca}
\alias{INFO_ppca}
\title{Information Criterion for PPCA.}
\usage{
INFO_ppca(lambda, M, verbose = FALSE, tau = 0.001, AIC = TRUE)
}
\arguments{
\item{lambda}{a numeric vector of positive sample eigenvalues pf length n}

\item{M}{a positive integer for the number of observations or features.}

\item{verbose}{a logical specifying whether the posterior evidence or
the integer that minimized the evidence should be returned}

\item{tau}{a tolerance threshold for the smallest eigenvalue, the default value is 0.001.}

\item{AIC}{a logical indicator to use AIC as the information criterion, if FALSE, then BIC is used; the default option is AIC.}
}
\value{
an integer K between 1 and n that minimizes the AIC or BIC.
}
\description{
The function returns the dimension that minimized the AIC or BIC based on the
   profile log-likelihood while considering all possible dimensions.
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale((X))))$mat))$val
INFO_ppca(lambda = eigen_values, M = 100)
INFO_ppca(lambda = eigen_values, M = 100, AIC = FALSE)
INFO_ppca(lambda = eigen_values, M = 5000)
INFO_ppca(lambda = eigen_values, M = 5000, AIC = FALSE)
}

}
\references{
Akaike, H. (1974), A new look at the statistical model identification, **IEEE Transactions on Automatic Control**, *19*(6): 716–723, <doi:10.1109/TAC.1974.1100705>
}
\keyword{criterion,}
\keyword{information}
\keyword{log-likelihood,}
\keyword{model}
\keyword{profile}
\keyword{selection}
