% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative.R
\name{Lawley.Test}
\alias{Lawley.Test}
\title{Likelihood Ratio Tests for equality of last $n-K$ eigenvalues.}
\usage{
Lawley.Test(lambda, M)
}
\arguments{
\item{lambda}{a numeric vector of positive sample eigenvalues that sums to $N$}

\item{M}{a positive integer for the number of observations or features``}
}
\value{
an integer $K$ between 1 and $N$ that minimizes the LRT $p$-value.
}
\description{
The function returns the dimension that minimizes the $p$-value from the
  likelihood ratio test (LRT) derived by Lawley (1956).
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(t(X))))$mat))$val
Lawley.Test(lambda = eigen_values, M = 1000)
}

}
\references{
Lawley, D. N. (1956) Tests of significance for the latent roots of
   stats::covariance and correlation matrices. \emph{Biometrika} \strong{43}.1/2: 128-136.
}
\author{
Wei Q. Deng, \email{deng@utstat.toronto.edu}
}
\keyword{hypothesis}
\keyword{likelihood}
\keyword{ratio}
\keyword{test,}
\keyword{testing,}
