% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative.R
\name{LawleyTest}
\alias{LawleyTest}
\title{Likelihood Ratio Tests for equality of last $n-K$ eigenvalues.}
\usage{
LawleyTest(lambda, M)
}
\arguments{
\item{lambda}{a numeric vector of positive sample eigenvalues that sums to $n$}

\item{M}{a positive integer for the number of observations or features}
}
\value{
an integer $K$ between 1 and $n$ that minimizes the LRT $p$-value.
}
\description{
The function returns the dimension that minimizes the $p$-value from the
  likelihood ratio test (LRT) derived by Lawley (1956).
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(1000, mu = rep(0,10), Sigma = diag(c(5,4,2,4,1,1,1,1,1,1)))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(t(X))))$mat))$val
LawleyTest(lambda = eigen_values)
}


}
\references{
Lawley, D. N. (1956) Tests of significance for the latent roots of covariance and correlation matrices. **Biometrika** *43*.1/2: 128-136. <doi:10.2307/2333586>
}
\keyword{hypothesis}
\keyword{likelihood}
\keyword{ratio}
\keyword{test,}
\keyword{testing,}
