% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPCA_profile_log_likelihood.R
\name{loglk}
\alias{loglk}
\title{Profile log-likelihood of the PPCA model with sample eigenvalues.}
\usage{
loglk(lam, n, tau = 1e-05)
}
\arguments{
\item{lam}{a numerical vector of positive sample eigenvalues}

\item{n}{the full dimension}

\item{tau}{a tolerance threshold for the smallest eigenvalue, the default value is 0.001.}
}
\value{
profile log-likelihood of length \eqn{n-1}.
}
\description{
The function returns the profile log-likelihood
   of the PPCA model for each possible choice of
   \eqn{K (1, 2, \dots, n-1)} at their respective MLEs.
   The maximum choice was set at \eqn{n-1} because when \eqn{K=n},
   the profile log-likelihood is equal to that at \eqn{K=n-1}.
}
