% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minka.R
\name{minka2001}
\alias{minka2001}
\title{Automatic Choice of dimensionality for PCA using Laplace's method}
\usage{
minka2001(x = NULL, lambda = NULL, M = NULL, evidence = FALSE)
}
\arguments{
\item{x}{a data matrix with the number of rows to be reduced; only complete columns are used.}

\item{lambda}{a numeric vector of sample eigenvalues of the covariance matrix of t(\code{x})}

\item{M}{if \code{x} were not supplied, \code{M} should be given as the number of columns of \code{x}.}

\item{evidence}{a logical specifying whether the BIC evidence or
the integer that minimized the evidence should be returned}
}
\value{
an integer K between 1 and N that maximizes the posterior BIC approximated by Laplace's method.
}
\description{
The function returns the choice dimension for PCA under the PPCA setup using Laplace approximation.
}
\examples{
\dontrun{
X <- MASS::mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(X)))$mat))$val
minka2001(lambda = eigen_values, M = 100)
minka2001(lambda = eigen_values, M = 5000)
}

}
\references{
Minka, T. P. (2000). Automatic choice of dimensionality for PCA.
In \emph{NIPS} (Vol. \strong{13}, pp. 598-604).
}
\keyword{Bayesian}
\keyword{Laplace's}
\keyword{criterion,}
\keyword{evidence}
\keyword{information}
\keyword{log-likelihood,}
\keyword{method,}
\keyword{model}
\keyword{profile}
\keyword{selection,}
