% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passemier2017.R
\name{passemier}
\alias{passemier}
\title{A bias corrected criterion for selecting number of principal components}
\usage{
passemier(lambda = NULL, M = NULL, tau = 0.001, constant = 0.05)
}
\arguments{
\item{lambda}{a numeric vector of sample eigenvalues of length $n$.}

\item{M}{the number of observations.}

\item{tau}{a tolerance threshold for the smallest eigenvalue, the default value is 0.001.}

\item{constant}{a small prefixed constant and set to the recommended value of 0.05. See Passemier et al., (2017) for details.}
}
\value{
an integer K
}
\description{
The function returns the choice for PCA as a by-product of the bias-corrected residual variance estimate.
}
\examples{
\dontrun{
X <- MASS::mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(X)))$mat))$val
passemier(lambda = eigen_values, M = 100)
passemier(lambda = eigen_values, M = 5000)
}

}
\references{
Passemier, D., Li, Z., & Yao, J. (2017). On estimation of the noise variance in high dimensional probabilistic principal component analysis. **Journal of the Royal Statistical Society: Series B (Statistical Methodology)**, *79*(1), 51-67. <doi:10.1111/rssb.12153>
}
\keyword{Bayesian}
\keyword{Laplace's}
\keyword{criterion,}
\keyword{evidence}
\keyword{information}
\keyword{log-likelihood,}
\keyword{method,}
\keyword{model}
\keyword{profile}
\keyword{selection,}
