% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPPCA.R
\name{ppca_log_penal}
\alias{ppca_log_penal}
\title{Penalized profile log-likelihood of a PPPCA model}
\usage{
ppca_log_penal(delta, lambda, lk = T)
}
\arguments{
\item{delta}{the value of the tuning parameter, must be a positive real number}

\item{lambda}{a numerical vector of positive sample eigenvalues}

\item{lk}{a logical specifying whether the penalized profile log-likelihood or
the integer that maximizes the penalized profile log-likelihood should be returned.}
}
\value{
an integer K that maximizes the penalized profile log-likelihood for the given \code{delta} value.
}
\description{
The function returns either the penalized profile
    log-likelihood or the value that maximizes the penalized
    profile log-likelihood for a given tuning parameter value.
}
\note{
This is an internal function called only by \code{pPPCA()}.
}
\author{
Wei Q. Deng, \email{deng@utstat.toronto.edu}
}
\keyword{PCA,}
\keyword{log-likelihood,}
\keyword{penalized}
\keyword{penalty}
\keyword{probabilistic}
\keyword{profile}
