% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative.R
\name{ppcaCV}
\alias{ppcaCV}
\title{Cross-validation for the PPCA model.}
\usage{
ppcaCV(k = NULL, x = NULL, fold = 5)
}
\arguments{
\item{k}{an integer indicating the tested dimension, should be between 1 and the number of rows of \code{x}.}

\item{x}{a data matrix with the number of rows to be reduced; only complete columns are used.}

\item{fold}{an integer indicating the number of folds to be used.}
}
\value{
profile log-likelihood of the remaining fold evaluated at MLEs computed from other folds.
}
\description{
The function returns the profile log-likelihood
   of the PPCA model at respective MLEs for a specific
   choice of $K=k$ evaluated by cross-validation.
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(1000, mu = rep(0,50), Sigma = diag(1,50))
ppcaCV(k = 5, x = X, fold = 5) # 5-fold cross-validation.
}

}
\references{
Tipping, M. E., and Bishop, C. M. (1999). Probabilistic principal component analysis.
   \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
   \strong{61}(3), 611-622.
}
\author{
Wei Q. Deng, \email{deng@utstat.toronto.edu}
}
\keyword{PCA,}
\keyword{cross}
\keyword{log-likelihood,}
\keyword{model}
\keyword{probabilistic}
\keyword{profile}
\keyword{selection}
\keyword{validation,}
