% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPCA_profile_log_likelihood.R
\name{ppcaLog}
\alias{ppcaLog}
\title{Profile log-likelihood of the PPCA model.}
\usage{
ppcaLog(x = NULL, lambda = NULL, M = NULL, param = NULL, EM = FALSE)
}
\arguments{
\item{x}{a data matrix where the number of rows is to be reduced; only complete columns are used}

\item{lambda}{a numeric vector of sample eigenvalues of the covariance matrix of t(\code{x})}

\item{M}{if \code{x} were not supplied, \code{M} should be given as the number of columns of \code{x}.}

\item{param}{a list of MLEs to be supplied if the profile log-likelihood was to
be evaluate at external MLES; the first list should contain a loading
matrix \eqn{W} with dimension \eqn{n} by \eqn{K} (\eqn{\le n});
and the second a numeric value between 0 and 1 for the
residual variance \code{sigma2}.}

\item{EM}{a logic indicator for whether the profile log-likelihood should be computed for
the MLE estimated using EM algorithms, if FALSE, the profile log-likelihood will be
evaluated by substitute the analytical formulation of MLEs.
Note that if \eqn{n} > \eqn{m}, it will not be possible to use this option.}
}
\value{
profile log-likelihood of length \eqn{n-1}.
}
\description{
The function returns the profile log-likelihood
   of the PPCA model for each possible choice of
   \eqn{K (1, 2, \dots, n-1)} at their respective MLEs.
   The maximum choice was set at \eqn{n-1} because when \eqn{K=n},
   the profile log-likelihood is equal to that at \eqn{K=n-1}.
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(X)))$mat))$val
U <- pracma::randortho(10)
W <- U\%*\%diag(1, 10)\%*\%t(U)
sigma2 <- 0.4
ppcaLog(x = t(X), EM=TRUE) # supply a data matrix
ppcaLog(x = t(X), param = list(W, sigma2)) # supply a data matrix and MLEs
ppcaLog(lambda = eigen_values, M = 1000) # supply the sample eigenvalues
}
}
\references{
Tipping, M. E., and Bishop, C. M. (1999). Probabilistic principal component analysis.
   \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, \strong{61}(3), 611-622.
}
\author{
Wei Q. Deng, \email{utstat.toronto.edu}
}
\keyword{Estimates,}
\keyword{Expectation}
\keyword{Likelihood}
\keyword{Maximization,}
\keyword{Maximum}
\keyword{PCA,}
\keyword{and}
\keyword{log-likelihood}
\keyword{probabilistic}
\keyword{profile}
