% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPPCA.R
\name{pPPCA_INFO}
\alias{pPPCA_INFO}
\title{Penalized profile log-likelihood of a PPPCA model}
\usage{
pPPCA_INFO(lam, Tvotes = 100, M, AIC = TRUE, tau = 0.001, beta = NULL)
}
\arguments{
\item{lam}{a numerical vector of positive sample eigenvalues}

\item{Tvotes}{the number of possible tuning parameter values to be searched}

\item{M}{the number of observations used to calculate sample covariance.}

\item{AIC}{a logical indicator to use AIC as the information criterion, if FALSE, then BIC is used; the default option is AIC.}

\item{tau}{a tolerance threshold for the smallest eigenvalue, the default value is 0.001.}

\item{beta}{a numeric between 0 and 1 indicating the weight towards penalty function 1 or 2.}
}
\value{
an integer K that maximizes the penalized profile log-likelihood for the given \code{delta} value.
}
\description{
The function returns either the penalized profile
    log-likelihood or the value that maximizes the penalized
    profile log-likelihood for a given tuning parameter value.
}
\keyword{PCA,}
\keyword{log-likelihood,}
\keyword{penalized}
\keyword{penalty}
\keyword{probabilistic}
\keyword{profile}
