% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPPCA_profile_log_likelihood.R
\name{pppcaProfile}
\alias{pppcaProfile}
\title{Penalized profile log-likelihood of a PPPCA model}
\usage{
pppcaProfile(delta, x = NULL, lambda = NULL, M = NULL, lk = FALSE)
}
\arguments{
\item{delta}{the value of the tuning parameter, must be a positive real number}

\item{x}{a data matrix with the number of rows to be reduced; only complete columns are used.}

\item{lambda}{a numerical vector of positive sample eigenvalues}

\item{M}{if \code{x} were not supplied, \code{M} should be given as the number of columns of \code{x}.}

\item{lk}{a logical specifying whether the penalized profile log-likelihood or
the integer that maximizes the penalized profile log-likelihood should be returned.}
}
\value{
an integer K that maximizes the penalized profile log-likelihood for the given \code{delta} value.
}
\description{
The function returns either the penalized profile
    log-likelihood or the value that maximizes the penalized
    profile log-likelihood for a given tuning parameter value.
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(X)))$mat))$val
pppcaProfile(delta = 0, lambda = eigen_values, M = 1000, lk = FALSE)
pppcaProfile(delta = 0, lambda = eigen_values, M = 1000, lk = TRUE)
}
}
\author{
Wei Q. Deng, \email{deng@utstat.toronto.edu}
}
\keyword{PCA,}
\keyword{log-likelihood,}
\keyword{penalized}
\keyword{penalty}
\keyword{probabilistic}
\keyword{profile}
