% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zhu_Ghodsi.R
\name{ZG}
\alias{ZG}
\title{Automatic Dimensionality Selection from the Scree Plot via Profile Likelihood}
\usage{
ZG(lambda, tau = 0.001, verbose = FALSE)
}
\arguments{
\item{lambda}{a numerical vector of sample eigenvalues of length $n$}

\item{tau}{a tolerance threshold for the smallest eigenvalue, the default value is 0.001.}

\item{verbose}{a logical to indicate whether the details of the profile log-likelihood evidence should be shown}
}
\value{
an integer $K$ between 1 and $n$.
}
\description{
The function returns the estimated dimension from a scree plot using a profile
likelihood criterion. The method requires only the sample eigenvalues.
}
\examples{
\dontrun{
library(MASS)
normdata <- mvrnorm(1000, mu = rep(0,50), Sigma = diag(1,50))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(normdata)))$mat))$val

ZG(lambda = eigen_values) # supply the sample eigenvalues
}

}
\references{
Zhu, M., & Ghodsi, A. (2006). Automatic dimensionality selection from the scree plot via the use of profile likelihood. **Computational Statistics & Data Analysis**, *51*(2), 918-930. <doi:10.1016/j.csda.2005.09.010>
}
\author{
Wei Q. Deng, \email{dengwq@mcmaster.ca} and Radu V. Craiu \email{craiu@utoronto.ca}
}
