% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minka.R
\name{minka2001}
\alias{minka2001}
\title{Automatic Choice of Dimensionality for PCA using Bayesian Posterior}
\usage{
minka2001(lambda, M, verbose = FALSE, tau = 0.001, BIC = FALSE)
}
\arguments{
\item{lambda}{a numeric vector of sample eigenvalues of the covariance matrix of \code{X} of dimension n by M.}

\item{M}{the number of columns of \code{X}.}

\item{verbose}{a logical specifying whether the posterior evidence or
the integer that minimized the evidence should be returned}

\item{tau}{a tolerance threshold for the smallest eigenvalue, the default value is 0.001.}

\item{BIC}{a logical indicating whether the Laplace's method or the BIC approximation should be used.}
}
\value{
an integer K between 1 and n that maximizes the posterior evidence  by Laplace's method or BIC approximation.
}
\description{
The function returns the choice dimension for PCA under the PPCA setup using Laplace approximation.
}
\examples{
\dontrun{
X <- MASS::mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(X)))$mat))$val
minka2001(lambda = eigen_values, M = 100, BIC=TRUE)
minka2001(lambda = eigen_values, M = 100, BIC=FALSE)
minka2001(lambda = eigen_values, M = 5000)
}

}
\references{
Minka, T. (2000). Automatic choice of dimensionality for PCA. **Advances in neural information processing systems**, *13*, 598-604. [http://dblp.uni-trier.de/db/conf/nips/nips2000.html#Minka00]
}
\keyword{Bayesian}
\keyword{Laplace's}
\keyword{criterion,}
\keyword{evidence}
\keyword{information}
\keyword{log-likelihood,}
\keyword{method,}
\keyword{model}
\keyword{profile}
\keyword{selection,}
