% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPPCA.R
\name{pPPCA}
\alias{pPPCA}
\title{Penalized Probabilistic PCA}
\usage{
pPPCA(x = NULL, eigenvals = NULL, Tvotes = 5000, var_tol = 0.01,
  tol = 1e-06, maxits = 500, printComp = F, verbose = F)
}
\arguments{
\item{x}{a data matrix with the number of rows to be reduced; only complete columns are used.}

\item{eigenvals}{a numerical vector of sample eigenvalues}

\item{Tvotes}{the number of possible tuning parameter values to be searched}

\item{var_tol}{the tolerance on the empirical error variance to be greater than 0,
the default proportion is set to 0.01}

\item{tol}{a tolerance level for the EM algorithm to terminate computations.}

\item{maxits}{the maximum number of iterations in EM to find the principal components}

\item{printComp}{a logical to indicate whether only the principal components should be printed}

\item{verbose}{a logical to indicate whether the details of the penalized voting results should be shown}
}
\value{
an integer $K$ between 1 and $n$.
}
\description{
The function returns the results of penalized profile
   log-likelihood given a matrix of data or a vector of sample
   eigenvalues. The data matrix is assumed to follow the decomposition
   \eqn{X = WL + \epsilon}, where rows of \eqn{X} are decomposed to a linear projection
   in an orthogonal space plus error. The solution finds the
   rank of \eqn{W}, which represents some hidden structure in
   the data, such that \eqn{X-WL} have independent and
   identically distributed components.
}
\examples{
\dontrun{
library(MASS)
normdata <- mvrnorm(1000, mu = rep(0,50), Sigma = diag(1,50))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(normdata)))$mat))$val

pPPCA(x = t(normdata)) # supply the data matrix
pPPCA(eigenvals = eigen_values) # supply the sample eigenvalues
}

}
\references{
Tipping, M. E., and Bishop, C. M. (1999). Probabilistic principal component analysis.
  \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
  \strong{61}(3), 611-622.
}
\author{
Wei Q. Deng, \email{deng@utstat.toronto.edu}
}
\keyword{PCA,}
\keyword{dimension}
\keyword{effective}
\keyword{log-likelihood,}
\keyword{parameter,}
\keyword{penalized}
\keyword{penalty}
\keyword{probabilistic}
\keyword{profile}
\keyword{tuning}
